requirements.txt
albumentations==1.1.0
astropy==5.0.1
asttokens==2.0.5
backcall==0.2.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fonttools==4.29.1
imageio==2.16.1
ipykernel==6.9.1
ipython==8.1.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
networkx==2.7.1
numpy==1.22.2
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pyerfa==2.0.0.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
stack-data==0.2.0
threadpoolctl==3.1.0
tifffile==2022.2.9
torch==1.11.0+cpu
torchaudio==0.11.0+cpu
torchvision==0.12.0+cpu
tornado==6.1
traitlets==5.1.1
typing-extensions==4.1.1
wcwidth==0.2.5
